requirements.txt
amqp==5.1.0
asgiref==3.5.0
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
dj-rest-auth==2.2.3
Django==4.0.2
django-allauth==0.49.0
django-filter==21.1
django-ipware==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
factory-boy==3.2.1
Faker==13.3.1
idna==3.3
iniconfig==1.1.1
kombu==5.2.4
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
